requirements.txt
asgiref==3.7.2
Django==4.2.4
django-cors-headers==4.2.0
django-dotenv==1.4.2
djangorestframework==3.14.0
PyMySQL==1.1.0
python-dotenv==1.0.0
pytz==2023.3
sqlparse==0.4.4
